Title: 세 수의 합과 평균 구하기, Time: 135ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241119/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -8,6 +8,6 @@ public class Main {
         int b = sc.nextInt();
         int c = sc.nextInt();
         System.out.println(a+b+c);
-        System.out.printf("%.0f",(double)(a+b+c)/3);
+        System.out.printf("%d",(a+b+c)/3);
     }
 }
